+++ src/pages/signup/index.tsx
@@ -1,5 +1,5 @@
-import Button from "@/components/button/Button"
-import FloatingButton from "@/components/button/FloatingButton"
+import Button from "@/components/common/Button"
+import FloatingButton from "@/components/common/FloatingButton"
 import Input from "@/components/input/Input";
 
 export default function Signup() {
@@ -13,7 +13,7 @@ export default function Signup() {
     return (
         <div className="w-1/4">
             <Button option="solid" size="xsmall" text="테스트" disabled={false}/>
-            {/* <FloatingButton option='cancel' text='할 일 추가' disabled={false} /> */}
+            <FloatingButton option='cancel' text='할 일 추가' disabled={true} />
             {/* <FloatingButton option='add' text='할 일 추가' disabled={false} onClick={testEvent}/> */}
             {/* <FloatingButton option='success' text='할 일 추가' disabled={false} /> */}
             <Input labeltext="테스트" option='text' errortext={testerrer} placeholder='미리보기' ></Input>
