+++ src/types/ButtonProp.ts
@@ -2,5 +2,4 @@ export interface ButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElemen
     option : "solid" | "outlined" | "outlinedSecondary" | "danger";
     size : "large" | "xsmall";
     text : string;
-    disabled? : boolean;
 }
