+++ src/types/FloatingButtonProp.ts
@@ -1,5 +1,4 @@
 export interface FloatingButtonProp extends React.ButtonHTMLAttributes<HTMLButtonElement> {
     option : 'add' | 'success' | 'cancel';
     text : string;
-    disabled? : boolean;
 }
